Moved user models to the core folder
