+++ frontend/src/app/auth/components/restore-page/restore-page.component.ts
@@ -11,7 +11,7 @@ import { Subject } from 'rxjs';
 import { delay, takeUntil } from 'rxjs/operators';
 import { AuthService } from 'src/core/services/auth.service';
 import { ValidationConstants } from 'src/entity-models/const-resources/validation-constraints';
-import { UserResetPasswordModel } from 'src/entity-models/user-reset-password-model';
+import { UserResetPasswordModel } from 'src/core/models/user/user-reset-password-model';
 import { InputComponent } from 'src/shared/components/tasque-input/input.component';
 
 @Component({
@@ -156,7 +156,7 @@ export class RestorePageComponent implements OnInit, OnDestroy {
 
   submitPassword(): void {
 
-    if(!this.resetForm.valid){
+    if (!this.resetForm.valid) {
       this.resetForm.markAllAsTouched();
       return;
     }
