--- frontend/src/entity-models/user-model.ts
@@ -1,8 +1,0 @@
-export interface UserModel {
-    id: number,
-    name: string,
-    email: string,
-    password: string,
-    salt: string
-}
-
